standardize connector layout for snowflake
